Update interferometer.tex
I have checked all the calculations by I-Sheng, the only change was the factor of 0.5 missing in a second order term that I mentioned to I-Sheng. I also included the change from proper velocity to coordinate velocity, it turns out that since that will be a second order effect I only had to change one term.